requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pinecone-client>=2.2.1
transformers>=4.28.0
sentence-transformers>=2.2.2
torch>=2.0.0
torchvision>=0.15.0
openai-whisper>=20230308
python-multipart>=0.0.6
python-dotenv>=1.0.0
pytest>=7.0.0
